Implement flow-based storage. (#54)
* Add FlowStore without proof.

* Add merkle tree crate.

* Pass merkle tree compilation.

* Pass a simple test.

* Implement range proof.

* Allow building with intermediate nodes.

* Allow the merkle tree to have missing data.

* Implement compatible LogStore traits for flow storage.

* Add unit tests and pass tests without proofs.

* Fix anyhow version.

* Implement merkle tree maintainance for flow-storage.

* Pass unit tests.

* Fix more.

* Fix some compile errors.

* Add log sync for new flow-based contracts.

* Pass compilation. Segment proof generation and validation are incomplete.

* Pass example_test.py.

* Fix more test issues.

* Fix unit tests.

* Update contract and tests.

* Save.

* Pass example tests.

* Update contract.

* Update token contract.

* Fix clippy warnings.

* Fix more tests.

* Fix comments.

+++ node/chunk_pool/src/mem_pool.rs
@@ -167,7 +167,7 @@ impl Inner {
                 .get_or_insert_with(Default::default)
                 .push_back(ChunkArray {
                     data: segment,
-                    start_index: start_index as u32,
+                    start_index: start_index as u64,
                 });
             return Ok(Some((file.tx_seq, file.segments.take().unwrap())));
         }
@@ -199,7 +199,7 @@ impl Inner {
             .get_or_insert_with(Default::default)
             .push_back(ChunkArray {
                 data: segment,
-                start_index: start_index as u32,
+                start_index: start_index as u64,
             });
 
         Ok(None)
